// src/app/models/product.ts
export interface Product {
  id?: number; // 1
  description: string; // Daghestanais d'origine, je me suis entrainé avec Khabib Nurmagomedov.
  price: number; // 299.99
  username: string; // EnormeZboubDu92
  height: number; // 185
  weight: number; // 105
  ratio: string; // '10-0'
  thumbnail: string; // https://images.unsplash.com/photo-1491756975177-a13d74ed1e2f?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MTR8fGZpZ2h0ZXJ8ZW58MHx8MHx8fDA%3D
  active: boolean;
}